+++ R/fitting.R
@@ -11,7 +11,7 @@
 #' @param nstudies Numeric, number of starting values in the LHS.
 #' @param initial_pop_constraints Boolean, whether to use additional constraints when sampling initial parameter values.
 #' @param error_function Character, error function to be used when fitting the model.
-#' @param model_version Character, version of the model ("TMM" or "PIM").
+#' @param model_version Character, threshold modulation model ("TMM3" or "TMM4") or parallel inhibition model ("PIM").
 #' @param method Character, optimisation method (DEoptim seems to work best).
 #' @param maxit Numeric, maximum number of iterations.
 #' @param verbose Boolean, whether to print progress during fitting.
@@ -22,7 +22,7 @@
 #'
 #' @examples
 #' \dontrun{
-#' plausible "true" parameter values in the TMM
+#' plausible "true" parameter values in the TMM (with 4 free parameters)
 #' true_pars <- c(1.1, 0.5, 0.3, 1.25)
 #'
 #' # simulating data using these parameter values
@@ -31,7 +31,7 @@
 #'     nsamples = 2000,
 #'     true_pars = true_pars,
 #'     action_mode = "imagined",
-#'     model_version = "TMM"
+#'     model_version = "TMM4"
 #'     )
 #'
 #' # fitting the model
@@ -40,7 +40,7 @@
 #'     nsims = 200,
 #'     error_function = "g2",
 #'     method = "DEoptim",
-#'     model_version = "TMM",
+#'     model_version = "TMM4",
 #'     lower_bounds = c(0, 0.25, 0.1, 1),
 #'     upper_bounds = c(2, 1.25, 0.6, 2),
 #'     initial_pop_constraints = TRUE,
@@ -64,7 +64,7 @@ fitting <- function (
         nstudies = 200,
         initial_pop_constraints = FALSE,
         error_function = c("g2", "rmse", "sse", "wsse", "ks"),
-        model_version = c("TMM", "PIM"),
+        model_version = c("TMM3", "TMM4", "PIM"),
         method = c(
             "SANN", "GenSA", "pso", "hydroPSO", "DEoptim",
             "Nelder-Mead", "BFGS", "L-BFGS-B", "bobyqa", "nlminb",
@@ -76,7 +76,11 @@ fitting <- function (
     # defining parameter names according to the chosen model (if null)
     if (is.null(par_names) ) {
 
-        if (model_version == "TMM") {
+        if (model_version == "TMM3") {
+
+            par_names <- c("exec_threshold", "peak_time_activ", "curvature_activ")
+
+         } else if (model_version == "TMM4") {
 
             par_names <- c("amplitude_activ", "peak_time_activ", "curvature_activ", "exec_threshold")
 
@@ -93,10 +97,10 @@ fitting <- function (
     stopifnot("nsims must be a numeric..." = is.numeric(nsims) )
     stopifnot("nstudies must be a numeric..." = is.numeric(nstudies) )
 
-    # testing whether only 4 pars have been specified
-    stopifnot("par_names must be a numeric of length 4..." = length(par_names) == 4)
-    stopifnot("lower_bounds must be a numeric of length 4..." = length(lower_bounds) == 4)
-    stopifnot("upper_bounds must be a numeric of length 4..." = length(upper_bounds) == 4)
+    # testing whether only 3 or 4 pars have been specified
+    stopifnot("par_names must be a numeric of length 3 or 4..." = length(par_names) %in% c(3, 4) )
+    stopifnot("lower_bounds must be a numeric of length 3 or 4..." = length(lower_bounds) %in% c(3, 4) )
+    stopifnot("upper_bounds must be a numeric of length 3 or 4..." = length(upper_bounds) %in% c(3, 4) )
 
     # method should be one of above
     method <- match.arg(method)
@@ -222,7 +226,7 @@ fitting <- function (
                 # F is the mutation constant (defaults to 0.8)
                 F = 0.9,
                 # crossover probability (recombination) (defaults to 0.5)
-                CR = 0.95,
+                CR = 0.9,
                 # c controls the speed of the crossover adaptation
                 # when strategy = 6 (defaults to 0)
                 # c = 0.1,
@@ -314,7 +318,7 @@ plot.DEoptim_momimi <- function (
         x, original_data,
         method = c("ppc", "latent"),
         action_mode = c("executed", "imagined"),
-        model_version = c("TMM", "PIM"),
+        model_version = c("TMM3", "TMM4", "PIM"),
         ...
         ) {
 
@@ -356,7 +360,70 @@ plot.DEoptim_momimi <- function (
 
     } else if (method == "latent") {
 
-        if (model_version == "TMM") {
+        if (model_version == "TMM3") {
+
+            par_names <- c("exec_threshold", "peak_time_activ", "curvature_activ")
+
+            parameters_estimates_summary <- paste(as.vector(rbind(
+                paste0(par_names, ": "),
+                paste0(as.character(round(estimated_pars, 3) ), "\n")
+                ) ), collapse = "") %>% stringr::str_sub(end = -2)
+
+            model(
+                nsims = 500, nsamples = 2000,
+                exec_threshold = estimated_pars[1] * 1.5,
+                imag_threshold = 0.5 * estimated_pars[1] * 1.5,
+                amplitude_activ = 1.5,
+                peak_time_activ = log(estimated_pars[2]),
+                curvature_activ = estimated_pars[3],
+                model_version = model_version,
+                full_output = TRUE
+                ) %>%
+                tidyr::pivot_longer(cols = .data$activation) %>%
+                ggplot2::ggplot(
+                    ggplot2::aes(
+                        x = .data$time, y = .data$value,
+                        group = interaction(.data$sim, .data$name)
+                        )
+                    ) +
+                # plotting the motor execution and motor imagery thresholds
+                geomtextpath::geom_labelhline(
+                    yintercept = estimated_pars[1] * 1.5,
+                    linetype = 2,
+                    hjust = 0.6,
+                    label = "Motor execution threshold"
+                    ) +
+                geomtextpath::geom_labelhline(
+                    yintercept = 0.5 * estimated_pars[1] * 1.5,
+                    linetype = 2,
+                    hjust = 0.6,
+                    label = "Motor imagery threshold"
+                    ) +
+                # plotting average
+                ggplot2::stat_summary(
+                    ggplot2::aes(group = .data$name),
+                    fun = "median", geom = "line",
+                    linewidth = 1, alpha = 1,
+                    show.legend = FALSE
+                    ) +
+                # displaying estimated parameter values
+                ggplot2::annotate(
+                    geom = "label",
+                    x = Inf, y = Inf,
+                    hjust = 1, vjust = 1,
+                    label = parameters_estimates_summary,
+                    family = "Courier"
+                    ) +
+                ggplot2::theme_bw(base_size = 12, base_family = "Open Sans") +
+                ggplot2::labs(
+                    title = "Latent function",
+                    x = "Time within a trial (in seconds)",
+                    y = "Activation/inhibition (a.u.)",
+                    colour = "",
+                    fill = ""
+                    )
+
+        } else if (model_version == "TMM4") {
 
             par_names <- c("amplitude_activ", "peak_time_activ", "curvature_activ", "exec_threshold")
 
@@ -372,7 +439,7 @@ plot.DEoptim_momimi <- function (
                 amplitude_activ = estimated_pars[1],
                 peak_time_activ = log(estimated_pars[2]),
                 curvature_activ = estimated_pars[3],
-                model_version = "TMM",
+                model_version = model_version,
                 full_output = TRUE
                 ) %>%
                 tidyr::pivot_longer(cols = .data$activation) %>%
@@ -382,13 +449,18 @@ plot.DEoptim_momimi <- function (
                         group = interaction(.data$sim, .data$name)
                         )
                     ) +
-                ggplot2::geom_hline(
+                # plotting the motor execution and motor imagery thresholds
+                geomtextpath::geom_labelhline(
                     yintercept = estimated_pars[4] * estimated_pars[1],
-                    linetype = 2
+                    linetype = 2,
+                    hjust = 0.6,
+                    label = "Motor execution threshold"
                     ) +
-                ggplot2::geom_hline(
+                geomtextpath::geom_labelhline(
                     yintercept = 0.5 * estimated_pars[4] * estimated_pars[1],
-                    linetype = 2
+                    linetype = 2,
+                    hjust = 0.6,
+                    label = "Motor imagery threshold"
                     ) +
                 # plotting average
                 ggplot2::stat_summary(
@@ -402,7 +474,7 @@ plot.DEoptim_momimi <- function (
                 ggplot2::annotate(
                     geom = "label",
                     x = Inf, y = Inf,
-                    hjust = 1, vjust = 1,
+                    hjust = 0, vjust = 1,
                     label = parameters_estimates_summary,
                     family = "Courier"
                     ) +